requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
cryptography
pandas
openpyxl
email-validator
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary